mini-run,python: move JobSpec class into main python bindings
